refactor: extract import functionality to separate module
