+++ crates/cli/commands/src/lib.rs
@@ -15,6 +15,7 @@ pub mod download;
 pub mod dump_genesis;
 pub mod import;
 pub mod import_era;
+pub mod import_op;
 pub mod init_cmd;
 pub mod init_state;
 pub mod launcher;
